// frontend/src/client/models/UserUpdate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UserUpdate = {
    username: string;
    password?: string;
    email?: string;
    mobile?: string;
    user_photo?: string;
    first_name?: string;
    last_name?: string;
    gender?: string;
    date_of_birth?: string;
    player_id?: number;
    coach_id?: number;
    referee_id?: number;
    manager_id?: number;
    disabled?: boolean;
    mobile_validation?: boolean;
    email_validation?: boolean;
    is_admin?: boolean;
    is_superuser?: boolean;
};